演算子関連のリファクタリング
