+++ Calculator/RPNComponents/FlowControl/End.cs
@@ -2,7 +2,7 @@
 {
     internal class End : FlowControlBase<End>
     {
-        public override string Name => ";";
+        protected override string Name => ";";
 
         public override void Execute(IRPNStack calculationTargets) { }
     }
