+++ Calculator/RPNComponents/ICalculationTarget.cs
@@ -1,5 +1,4 @@
-﻿using Calculator.RPNException;
-using System.Diagnostics.CodeAnalysis;
+﻿using System.Diagnostics.CodeAnalysis;
 
 namespace Calculator.RPNComponents
 {
@@ -11,7 +10,7 @@ namespace Calculator.RPNComponents
         /// <summary>
         /// スタックに積むことができる要素の型定義を表現したクラスかどうかを表す
         /// </summary>
-        bool IsDefinitionInstance { get; set; }
+        bool IsDefinitionInstance { get; init; }
 
         /// <summary>
         /// 指定したスタックに対して処理を行う
@@ -33,75 +32,5 @@ namespace Calculator.RPNComponents
         /// <param name="result"></param>
         /// <returns></returns>
         bool TryParse(string token, [NotNullWhen(true)] out ICalculationTarget result);
-    }
-
-    /// <summary>
-    /// 加減乗除の基本的な演算を扱う抽象クラス
-    /// </summary>
-    internal abstract class BasicOperator : ICalculationTarget
-    {
-        /// <summary>
-        /// 定義のみを扱うインスタンスか、実際に計算を実行するインスタンスかを判別する
-        /// </summary>
-        public bool IsDefinitionInstance { get; set; } = false;
-
-        /// <summary>
-        /// 指定したスタックから2つ値を取り出し、
-        /// </summary>
-        /// <param name="calculationTargets"></param>
-        /// <param name="calcFunc"></param>
-        /// <exception cref="Exception"></exception>
-        public void Execute(IRPNStack calculationTargets, Func<NumberTarget, NumberTarget, NumberModel> calcFunc)
-        {
-            if (IsDefinitionInstance) throw new RuntimeException("式が不正です");
-
-            var numberTarget1 = GetTwoNumberFromStack(calculationTargets);
-            var numberTarget2 = GetTwoNumberFromStack(calculationTargets);
-
-            var basicCalcResult = calcFunc(numberTarget1, numberTarget2);
-
-            var greatestCommonDivisor = CalculationHelper.CalcGreatestCommonDivisor(
-                basicCalcResult.Denominator, basicCalcResult.Numerator);
-
-            if (greatestCommonDivisor != 0)
-            {
-                basicCalcResult.Denominator /= greatestCommonDivisor;
-                basicCalcResult.Numerator /= greatestCommonDivisor;
-            }
-
-            calculationTargets.Push(
-                new NumberTarget(basicCalcResult.Denominator, basicCalcResult.Numerator));
-        }
-
-        protected static NumberTarget GetTwoNumberFromStack(IRPNStack calculationTargets)
-        {
-            var arg1 = calculationTargets.Pop();
-            if (arg1 is not NumberTarget numberTarget1)
-            {
-                arg1.Execute(calculationTargets);
-
-                arg1 = calculationTargets.Pop();
-
-                if (arg1 is not NumberTarget)
-                    throw new RuntimeException("式が不正です");
-
-                numberTarget1 = (NumberTarget)arg1;
-            }
-
-            return numberTarget1;
-        }
-
-        public abstract string Display();
-
-        public abstract void Execute(IRPNStack calculationTargets);
-
-        public abstract bool TryParse(string token, [NotNullWhen(true)] out ICalculationTarget result);
-
-        public class NumberModel
-        {
-            internal double Numerator { get; set; }
-
-            internal double Denominator { get; set; }
-        }
     }
 }
