+++ Calculator/RPNComponents/LogicalOperator/LessThan.cs
@@ -1,10 +1,4 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
-
-namespace Calculator.RPNComponents.LogicalOperator
+﻿namespace Calculator.RPNComponents.LogicalOperator
 {
     internal class LessThan : LogicalOperatorBase
     {
